refactor: restructured components
