+++ src/routes/app/login.svelte
@@ -1,0 +1,1 @@
+login
